+++ device-api/CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [SW-16508] Adding P2P DMA read/write commands and responses
 ### Changed
+- [SW-16508b] Split struct p2pdma_node into read/write node structs
+ * Positioning `struct p2pdma_(read/write)_node` fields to be consistent with struct `dma_(read/write)_node` for future possibility of merging the structs.
+ * Positioning fields in `struct p2pdma_(read/write)_node` such that peer device is always at same offset, so that the driver can translate the peer device address without identifying the command as read/write.
 ### Deprecated
 ### Removed
 ### Fixed